refactor: model import 수정
