+++ src/models/Concert.ts
@@ -11,8 +11,7 @@ import {
   CreatedAt,
   UpdatedAt,
 } from "sequelize-typescript";
-import User from "./User";
-import Post from "./Post";
+import { User, Post } from ".";
 
 @Table({
   tableName: "Concert",
