+++ src/models/User.ts
@@ -12,12 +12,7 @@ import {
   HasMany,
   HasOne,
 } from "sequelize-typescript";
-import Comment from "./Comment";
-import Like from "./Like";
-import Post from "./Post";
-import Scrap from "./Scrap";
-import UserInfo from "./UserInfo";
-import UserInterest from "./UserInterest";
+import { Comment, UserInfo, Like, Scrap, Post, UserInterest } from ".";
 
 @Table({
   tableName: "User",
